Add files via upload
KNN :
In KNN, I have set K = 6. The algorithm finds distance of a datapoint in test matrix from all points in train matrix, then checks the labels of the 6 nearest datapoints. It then assigns that label to the test datapoint whose number is highest in the selected 6 datapoints.
The accuracy is 0.8366781539293195, or around 83 %